+++ R/query_device.R
@@ -6,7 +6,9 @@
 #' @return a data-frame with a `device_id` column and the `$data` turned into
 #'    as many columns as expected
 #' @export
-#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc
+#' @importFrom httr2 request response req_perform resp_is_error
+#' @importFrom httr2 resp_body_json resp_status resp_status_desc
+#' @importFrom purrr possibly
 #'
 #' @examples
 #' \dontrun{
@@ -16,7 +18,7 @@ query_ap_device <- function(device_ip, query) {
   check_device_ip(device_ip)
   url <- glue::glue("http://{device_ip}:8050/{query}")
   req <- request(url)
-  resp <- req |> req_perform()
+  resp <- possibly(req |> req_perform(), otherwise = response(504))
   if (resp_is_error(resp)) {
     cli::cli_abort(c("Connection to device {.var device_ip} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
@@ -34,8 +36,9 @@ query_ap_device <- function(device_ip, query) {
 #' @return a data-frame with a row for each `device_id`, and the `$data` turned into
 #'    as many columns as expected
 #' @export
-#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc
-#' @importFrom purrr map map_lgl map_dfr
+#' @importFrom httr2 request response req_perform resp_is_error
+#' @importFrom httr2 resp_body_json resp_status resp_status_desc
+#' @importFrom purrr map map_lgl map_dfr possibly
 #'
 #' @examples
 #' \dontrun{
@@ -45,14 +48,16 @@ query_ap_device <- function(device_ip, query) {
 #' }
 query_ap_devices <- function(device_ip, query) {
   url <- glue::glue("http://{unique(device_ip)}:8050/{query}")
-  resp <- map(url, ~.x |> request() |> req_perform(error_call = rlang::caller_env()))
+  resp <- map(url, possibly(~.x |> request() |> req_perform(error_call = rlang::caller_env()),
+                            otherwise = response(504))
+  )
   response_is_error <- map_lgl(resp, resp_is_error)
   if (any(response_is_error)) {
-    cli::cli_abort(c("Connection to device {.var device_ip[response_is_error]} raise an error : ",
+    cli::cli_warn(c("Connection to device {.var device_ip[response_is_error]} raise an error : ",
                      "{resp_status(resp)} {resp_status_desc(resp)}."))
 
   } else {
-    info_lst <- map(resp, ~.x |> resp_body_json())
+    info_lst <- map(resp[!response_is_error], ~.x |> resp_body_json())
     map_dfr(info_lst, ~cbind(device_id = .x$deviceId, as.data.frame(.x$data))
     )
   }
@@ -73,7 +78,8 @@ query_ap_devices <- function(device_ip, query) {
 #' @return a data-frame with a `device_id` column and the `$data` turned into
 #'    as many columns as expected
 #' @export
-#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc
+#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc req_auth_basic response
+#' @importFrom purrr possibly
 #'
 #' @examples
 #' \dontrun{
@@ -114,11 +120,12 @@ query_enphase_device <- function(device_ip = "enphase.local", query, username = 
 #' @return a data-frame with a `device_id` column and the `$Body$Data` turned into
 #'    as many columns as expected
 #' @export
-#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc
+#' @importFrom httr2 request req_perform resp_is_error resp_body_json resp_status resp_status_desc req_auth_basic response
+#' @importFrom purrr possibly
 #'
 #' @examples
 #' \dontrun{
-#' query_fronius_device(device_ip = "192.168.0.234", query = "GetInverterRealtimeData.cgi?Scope=System")
+#' query_fronius_device(query = "GetInverterRealtimeData.cgi?Scope=System")
 #' }
 query_fronius_device <- function(device_ip = "fronius.local", query, username = Sys.getenv("FRONIUS_USERNAME"), password = Sys.getenv("FRONIUS_PASSWORD")) {
   check_device_ip(device_ip)
@@ -143,6 +150,6 @@ query_fronius_device <- function(device_ip = "fronius.local", query, username = 
 check_device_ip <- function(device_ip) {
   stopifnot("device_IP shall be an atomic character string" = length(device_ip) == 1)
   stopifnot("device_IP shall be of a minimal character length" = nchar(device_ip) >= 3)
-
-
+  # TODO minimal IP validation device_IP shall contain at least a  \. or at least two \:
+  # TODO use a proper IP address validation function
 }